+++ src/onegov.core/onegov/core/utils.py
@@ -1,3 +1,4 @@
+import bleach
 import hashlib
 import inspect
 import json
@@ -146,3 +147,60 @@ def groupbylist(*args, **kwargs):
 
     """
     return [(k, list(g)) for k, g in groupby(*args, **kwargs)]
+
+
+def sanitize_html(html):
+    """ Takes the given html and strips all but a whitelisted number of tags
+    from it.
+
+    """
+
+    if not html:
+        return html
+
+    allowed_tags = [
+        'a',
+        'abbr',
+        'b',
+        'br',
+        'blockquote',
+        'code',
+        'del',
+        'div',
+        'em',
+        'i',
+        'img',
+        'hr',
+        'li',
+        'ol',
+        'p',
+        'strong',
+        'sup',
+        'ul',
+        'h1',
+        'h2',
+        'h3',
+        'h4',
+        'h5',
+        'h6'
+    ]
+
+    allowed_attributes = {
+        'a': ['href', 'title'],
+        'abbr': ['title'],
+        'acronym': ['title'],
+        'img': ['src', 'alt', 'title']
+    }
+
+    return bleach.clean(html, tags=allowed_tags, attributes=allowed_attributes)
+
+
+def linkify(text):
+    """ Takes plain text and injects html links for urls and email addresses.
+
+    """
+
+    if not text:
+        return text
+
+    return bleach.linkify(text, parse_email=True)